requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.2
python-multipart>=0.0.6
pydantic>=2.3.0
langchain>=0.0.267
langchain_google_genai>=0.0.1
google-generativeai>=0.3.0
openai>=0.27.8
pypdf>=3.15.1
pytest>=7.4.0
httpx>=0.24.1
python-dotenv>=1.0.0
langgraph>=0.0.15